Move all the base code into the new structure
Luckily, this was a pretty small directory
